refactor(characters): extract Characters.Store
First slice of splitting the 1400-line Characters module along its
concern seams (#134). Disk persistence shares no helpers with the rest
of the module, so it moves wholesale into Characters.Store. The facade
keeps the full @doc strings and defdelegates, following the pattern
established by Characters.Advancement, so the published Hex API and its
docs are unchanged.
